+++ R/endpoints-bulk.R
@@ -3,14 +3,21 @@
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-make_bulk_create_job_url <- function(api_type=c("Bulk 1.0", "Bulk 2.0")){
+make_bulk_create_job_url <- function(api_type=c("Bulk 1.0", "Bulk 2.0"), 
+                                     query_operation=FALSE){
   # ensure we are authenticated first so the url can be formed
   sf_auth_check()
   api_type <- match.arg(api_type)
   if(api_type == "Bulk 2.0"){
-    sprintf("%s/services/data/v%s/jobs/ingest",
-            salesforcer_state()$instance_url,
-            getOption("salesforcer.api_version"))    
+    if(query_operation){
+      sprintf("%s/services/data/v%s/jobs/query",
+              salesforcer_state()$instance_url,
+              getOption("salesforcer.api_version"))      
+    } else {
+      sprintf("%s/services/data/v%s/jobs/ingest",
+              salesforcer_state()$instance_url,
+              getOption("salesforcer.api_version"))
+    }
   } else {
     sprintf("%s/services/async/%s/job",
             salesforcer_state()$instance_url,
@@ -23,15 +30,24 @@ make_bulk_create_job_url <- function(api_type=c("Bulk 1.0", "Bulk 2.0")){
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-make_bulk_get_job_url <- function(job_id, api_type=c("Bulk 1.0", "Bulk 2.0")){
+make_bulk_get_job_url <- function(job_id, 
+                                  api_type=c("Bulk 1.0", "Bulk 2.0"), 
+                                  query_operation=NULL){
   # ensure we are authenticated first so the url can be formed
   sf_auth_check()
   api_type <- match.arg(api_type)
   if(api_type == "Bulk 2.0"){
-    sprintf("%s/services/data/v%s/jobs/ingest/%s",
-            salesforcer_state()$instance_url,
-            getOption("salesforcer.api_version"), 
-            job_id) 
+    if(!is.null(query_operation) && query_operation){
+      sprintf("%s/services/data/v%s/jobs/query/%s",
+              salesforcer_state()$instance_url,
+              getOption("salesforcer.api_version"), 
+              job_id)       
+    } else {
+      sprintf("%s/services/data/v%s/jobs/ingest/%s",
+              salesforcer_state()$instance_url,
+              getOption("salesforcer.api_version"), 
+              job_id)       
+    }
   } else {
     sprintf("%s/services/async/%s/job/%s",
             salesforcer_state()$instance_url,
@@ -210,11 +226,19 @@ make_bulk_batches_url <- function(job_id, api_type=c("Bulk 1.0", "Bulk 2.0")){
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-make_bulk_query_url <- function(job_id, api_type=c("Bulk 1.0")){
+make_bulk_query_url <- function(job_id=NULL, 
+                                api_type=c("Bulk 1.0", "Bulk 2.0")){
   # ensure we are authenticated first so the url can be formed
   sf_auth_check()
   api_type <- match.arg(api_type)
-  if(api_type == "Bulk 1.0"){  
+  if(api_type == "Bulk 2.0"){
+    sprintf("%s/services/data/v%s/jobs/query",
+            salesforcer_state()$instance_url,
+            getOption("salesforcer.api_version"))
+  } else{
+    if(is.null(job_id)){
+      stop("The `job_id` must be provided to create Bulk 1.0 query", call.=FALSE)
+    }
     sprintf("%s/services/async/%s/job/%s/batch",
             salesforcer_state()$instance_url,
             getOption("salesforcer.api_version"), 
@@ -261,11 +285,28 @@ make_bulk_batch_details_url <- function(job_id, batch_id, api_type=c("Bulk 1.0")
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-make_bulk_query_result_url <- function(job_id, batch_id, result_id, api_type=c("Bulk 1.0")){
+make_bulk_query_result_url <- function(job_id, 
+                                       batch_id=NULL, 
+                                       result_id=NULL, 
+                                       api_type=c("Bulk 1.0", "Bulk 2.0")){
   # ensure we are authenticated first so the url can be formed
   sf_auth_check()
   api_type <- match.arg(api_type)
-  if(api_type == "Bulk 1.0"){
+  if(is.null(job_id)){
+    stop("The `job_id` must be provided to get query results", call.=FALSE)
+  }
+  if(api_type == "Bulk 2.0"){
+    sprintf("%s/services/data/v%s/jobs/query/%s/results",
+            salesforcer_state()$instance_url,
+            getOption("salesforcer.api_version"), 
+            job_id)
+  } else{
+    if(is.null(batch_id)){
+      stop("The `batch_id` must be provided to get Bulk 1.0 query results", call.=FALSE)
+    }    
+    if(is.null(result_id)){
+      stop("The `result_id` must be provided to get Bulk 1.0 query results", call.=FALSE)
+    }        
     sprintf("%s/services/async/%s/job/%s/batch/%s/result/%s",
             salesforcer_state()$instance_url,
             getOption("salesforcer.api_version"),